requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
psycopg2-binary~=2.9.9
python-dotenv==1.0.1
django-axes~=6.1.0
django-crispy-forms
crispy-bootstrap5

